refactored group form to it's own component
